Actually respect the "--with-chisel" option
It turns out that you need to modify the build.scala in order to make
SBT actually use the local versions of Chisel.  The current autoconf
stuff wasn't doing that, which means that despite the fact that Chisel
was being built it wasn't actually being used.

This is still sub-optimal: it globally installs the Chisel version
you're using for Sodor.  I know nothing of SBT, so I'm just not going
to touch that... :).