Find the fesvr headers in a standard way
Originally (in other words, yesterday), sodor found the RISC-V tools
(fesvr was the only important one) by looking for them only in PREFIX,
which is a very non-standard way to do things.  Eric's recent patch
made them look in $RISCV, which while being the way that the rest of
the RISC-V tools do it is still fairly non-standard.  Both of these
approaches had the problem of failing with a cryptic error message at
compile time as opposed to failing when "./configure" was run.  This
patch changes this behavior to be more in line with what autoconf
normally does while maintaining backwards compatibility with both
methods.

Firstly, a new "--with-riscv" option has been added to "./configure".
If this option is passed then the RISC-V tools will be searched for in
the given directory.  If "--with-riscv" is NOT passed on the command
line, then configure will check if the RISCV environment variable is
set and if it is check for the RISC-V tools there.  If both of these
options are not set, then configure will search for the RISC-V tools
in PREFIX, which is always set (it defaults to "/usr/local").

Additionally, "./configure" now actually checks for the presence of
"fesvr/htif_pthread.h" when it is run and provides an error message if
that file does not exist.